Reset money to previous value before the save
So, if you let the timer run down to 6:00 A.M., you lose 10% of your cash, the game saves, the new day starts, and NightOwl resets your money to what it was at 5:59.  There are two problems with this.  First, if you quit to the title and reload, your money doesn't get reset again on load, so you lose the 10%.  Second, if you shipped anything the day before, the cash reset wipes out the cash gained from shipping items.  This fixes both.  The first is fixed by resetting the cash value during the before save event, the second is fixed by reading the cash penalty amount from the mailbox before the mailbox is cleared, and adding the value rather than setting the total.