Add config to backup saves after save, not before
Adds a config option "BackupAfterSave" that causes save files to be
backed up after saving instead of before saving. This allows the
backup saves directory to contain a copy of all save files, not just
the old ones. This also removes the extremely small risk of causing
a crash before saving and thus losing progress. Defaults to backing
up the save before saving, which is the original behaviour.

Bumps version to 1.13.0.